Integration of simple I/O into a stand alone test for P3
This commit uses the infrastructure of simple_io to load a model
state from an input file to initialize the set of required fields
for running P3.

The input file contains the model state for 24 SCM cases.

Many things have been hard-coded to make this work.  A big TODO item will be
to generalize the interface so that P3 can also be run in a standard mode
without the SCM case input.

Comparison of the output field for this simulation with the output using
standard E3SM is not BFB, but the differences are incredibly close, O(1e-5) RMSE.

This implies that the errors may be based on precision.  The next set of work
will focus on determining where the non-BFB results are coming from.