Initial implementation of SCORPIO in SCREAM.
Adds the necessary cmake files to build SCORPIO as an external codebase
using a similar strategy to the KOKKOS build.

Adds a simple "Hello World" interface between scream and scorpio which
will be built up in subsequent commits.

Adds a simple scream_pio test to the set of scream tests.