Calling p3_main C++ code.
Set up initializer to use the same values as the p3_run_and_cmp test.

run_impl in atmosphere_microphysics.cpp now determines the local
values that are needed by P3 but not needed in other processes.

TODO: Still does not work on GPU's.  Most likely because run_impl
is using straight "for" loops to populate the local variables.  Currently the
Kokkos::parallel_reduce loops are causing the gpu version to throw an error.
These will need to be fixed in the next commit.