SCREAM: upgrades to FieldIdentifier
The class now stores also exposes a long name, while 'name' is now
referring to a short and handy name. The long name does not contribute
to the id string, so there *could* be two instances of this class that
have all members equal except the long name, and the op== would return
true. The FieldRepository will be the class most in charge of making
sure that this doesn't happen.

Furthermore, the class now uses a sort-of pimpl idiom to store the data
(the data struct is actually not hidden completely (it is in the header
file). The reason is to limit duplication of the same data around,
since FieldIdentifier's are quite commonly stored around classes,
such as atm processes, remappers, I/O, and more.