SCREAM: better handle of GroupRequest with a relative
In particular, we need to handle a request formulated in terms of a
relative that is on a different grid. In that case, the relative request
might be empty on the current grid.

E.g. GR1 on grid1, contains (f1,f2). GR2 on grid2 is an alias of GR1.
If no atm proc register any field on grid2, then GR1 on grid2 would be
empty (recall that the FM on each grid is independent of the others).
So the AD needs to "import" GR1 from grid1 to grid2, to handle the
relationship between GR1 and GR2.