SCREAM: added exclusive fields/groups to atm process interface
Exclusive fields are supposed to be used by one atm proc only.
They are meant to be used for internal state of a proc, that is,
not something that is returned/outputed/observable to/by the
rest of the atmosphere model.

E.g., some atm procs need the copy of field X at the end of
the previous call of the atm proc run method, so that they
can back out a tendency for it. Such state is irrelevant
for the rest of the atm, but must be saved/loaded for exact
restarts.