Introduce structures to SPA to manage SPA data and interpolation parameters.
This commit also updates EKAT and fixes a hack from previous commits that
had to make some variables "Updated" to use the EKAT linear interpolation
routine.  The EKAT fix made it possible to return these variables to being
"Required".

The structures introduced in this commit should simplify the set of variables
that need to be defined in run_impl, and lay the ground work for a SPA main
function that takes the structures as input, similar to what is done in P3.