SCREAM: fix how field stack classes keep track of parents/children
It all stemmed from FieldHeader trying to call shared_from_this
inside the constructor (which always yields a null shared_ptr).
To overcome this, one needs to perform the parent/children setup
outside the ctor. On the other hand, it is safe to do the cleanup
from inside the dtor, provided that one only uses weak_from_this,
since the last shared_ptr owning this object has been destroyed
(which is why we are inside the dtor).

To make the code slimmer, both FieldHeader and FieldTracking rely
on the FamilyTracking utility class. Notice that FieldTracking
only needs parent/children relationships to allow the update
of children's time stamp when the parent's time stemp is updated.