SCREAM: start to work on allowing output on dynamics grid
In case the input field manager is defined on a non-ref grid,
the output manager builds a remapper from the grid to the
corresponding unique grid, and runs the remapper before
the output, at every run call.

The most important example is output on the dyn grid. In this
case, we use a PhysicsGLL grid to remap the out fields onto,
and then output the remapped fields. Notice that the remap
dyn->physGLL is always invertible, while dyn->pg2 is not.